+++ src/main/kotlin/com/yeoljeong/autovote/infrastructure/security/oauth2/CustomOAuth2UserService.kt
@@ -1,4 +1,4 @@
-package com.yeoljeong.autovote.infrastructure.security
+package com.yeoljeong.autovote.infrastructure.security.oauth2
 
 import com.yeoljeong.autovote.domain.user.User
 import com.yeoljeong.autovote.domain.user.UserAppender
